requirements.txt
google-cloud-asset>=3.25.0,<4.0.0
google-auth>=2.30.0,<3.0.0